+++ src/Parser/Block/BlockStart.php
@@ -34,6 +34,8 @@ final class BlockStart
     /** @psalm-readonly-allow-private-mutation */
     private bool $replaceActiveBlockParser = false;
 
+    private bool $isAborting = false;
+
     private function __construct(BlockContinueParserInterface ...$blockParsers)
     {
         $this->blockParsers = $blockParsers;
@@ -55,6 +57,14 @@ final class BlockStart
     public function isReplaceActiveBlockParser(): bool
     {
         return $this->replaceActiveBlockParser;
+    }
+
+    /**
+     * @internal
+     */
+    public function isAborting(): bool
+    {
+        return $this->isAborting;
     }
 
     /**
@@ -97,5 +107,18 @@ final class BlockStart
     public static function of(BlockContinueParserInterface ...$blockParsers): self
     {
         return new self(...$blockParsers);
+    }
+
+    /**
+     * Signal that the block parsing process should be aborted (no other block starts should be checked)
+     *
+     * @internal
+     */
+    public static function abort(): self
+    {
+        $ret             = new self();
+        $ret->isAborting = true;
+
+        return $ret;
     }
 }